WIP @params

+++ protocol/synthetix/contracts/interfaces/IIssueUSDModule.sol
@@ -7,28 +7,42 @@ pragma solidity ^0.8.0;
 interface IIssueUSDModule {
     /**
      * @notice Emitted when {sender} mints {amount} of snxUSD with the specified staking position.
+     * @param accountId The id of the account for which USD was emitted.
+     * @param poolId The id of the pool whose collateral was used to emit the USD.
+     * @param collateralType The address of the collateral that is backing up the emitted USD.
+     * @param amount The amount of USD emitted.
+     * @param sender The address that triggered the operation.
      */
     event UsdMinted(
         uint128 indexed accountId,
         uint128 indexed poolId,
         address collateralType,
-        uint amount,
+        uint256 amount,
         address indexed sender
     );
 
     /**
      * @notice Emitted when {sender} burns {amount} of snxUSD with the specified staking position.
+     * @param accountId The id of the account for which USD was burnt.
+     * @param poolId The id of the pool whose collateral was used to emit the USD.
+     * @param collateralType The address of the collateral that was backing up the emitted USD.
+     * @param amount The amount of USD burnt.
+     * @param sender The address that triggered the operation.
      */
     event UsdBurned(
         uint128 indexed accountId,
         uint128 indexed poolId,
         address collateralType,
-        uint amount,
+        uint256 amount,
         address indexed sender
     );
 
     /**
      * @notice Mints {amount} of snxUSD with the specified staking position.
+     * @param accountId The id of the account that is minting USD.
+     * @param poolId The id of the pool whose collateral will be used to back up the mint.
+     * @param collateralType The address of the collateral that will be used to back up the mint.
+     * @param amount The amount of USD to be minted.
      *
      * Requirements:
      *
@@ -41,11 +55,15 @@ interface IIssueUSDModule {
         uint128 accountId,
         uint128 poolId,
         address collateralType,
-        uint amount
+        uint256 amount
     ) external;
 
     /**
      * @notice Burns {amount} of snxUSD with the specified staking position.
+     * @param accountId The id of the account that is burning USD.
+     * @param poolId The id of the pool whose collateral was used to back up the USD.
+     * @param collateralType The address of the collateral that was used to back up the USD.
+     * @param amount The amount of USD to be burnt.
      *
      * Requirements:
      *
@@ -57,6 +75,6 @@ interface IIssueUSDModule {
         uint128 accountId,
         uint128 poolId,
         address collateralType,
-        uint amount
+        uint256 amount
     ) external;
 }
